Make mouse interactions configurable
Fixes #51
