+++ config/config.go
@@ -54,62 +54,6 @@ type Config struct {
 	LogDir     string `yaml:"-" toml:"-"`
 }
 
-// Client holds information regarding how Stimmtausch runs.
-type Client struct {
-	Syslog  Syslog
-	Logging Logging
-	UI      UI
-}
-
-// Syslog holds nformation regarding the logging generated by the program (as
-// opposed to the connections).
-type Syslog struct {
-
-	// Whether or not to show the system log in a pane in the UI.
-	ShowSyslog bool `yaml:"show_syslog" toml:"log_level"`
-
-	// Lowest log level to show by default. Options are:
-	// TRACE, DEBUG, INFO (default), WARNING, ERROR, CRITICAL
-	LogLevel string `yaml:"log_level" toml:"log_level"`
-}
-
-// Logging holds information regarding logging from connections.
-type Logging struct {
-
-	// The time format to use when logging times.
-	// TODO validate
-	TimeString string `yaml:"time_string" toml:"time_string"`
-
-	// Whether or not to log timestamps.
-	LogTimestamps bool `yaml:"log_timestamps" toml"log_timestamps"`
-
-	// Whether or not to keep logs of the connection after disconnect.
-	LogWorld bool `yaml:"log_world" toml:"log_world"`
-}
-
-// UI holds information regarding the user interface.
-type UI struct {
-
-	// How many lines of scrollback (data received) to keep in memory.
-	Scrollback int
-
-	// How many lines of history (data sent) to keep in memory.
-	History int
-
-	// Whether or not to use a unified history buffer for all
-	// connections, or one per.
-	// TODO
-	UnifiedHistoryBuffer bool `yaml:"unified_history_buffer" toml:"unified_history_buffer"`
-
-	// Whether or not to use Vim-style key bindings
-	// TODO
-	VimKeybindings bool `yaml:"vim_keybindings" toml:"vim_keybindings"`
-
-	// Indentation of wrapped lines.
-	IndentFirst      int `yaml:"indent_first" toml:"indent_first"`
-	IndentSubsequent int `yaml:"indent_subsequent" toml:"indent_subsequent"`
-}
-
 func (c *Config) FinalizeAndValidate() []error {
 	log.Debugf("finalizing and validating config")
 	var errs []error
